объеденила обработчики исключений в один
